refactor: Share code between flows for systems boundaries
I found it easy to think about this as the boundaries, and where we
traverse them. This isn't the final step in sharing, but is a good
start imo.

TOWARDS PLA-481
