+++ flows/deindex.py
@@ -1,32 +1,40 @@
 import asyncio
 import base64
-import contextlib
 import json
 import os
 import re
-import tempfile
 from collections import Counter
-from collections.abc import Generator
 from dataclasses import dataclass
-from io import BytesIO
 from pathlib import Path
-from typing import Any, ContextManager, TypeAlias, TypeVar
+from typing import Any
 
 import boto3
-import vespa.querybuilder as qb
 from cpr_sdk.models.search import Concept as VespaConcept
 from cpr_sdk.models.search import Passage as VespaPassage
-from cpr_sdk.s3 import _get_s3_keys_with_prefix, _s3_object_read_text
 from cpr_sdk.search_adaptors import VespaSearchAdapter
 from cpr_sdk.ssm import get_aws_ssm_param
 from prefect import flow
 from prefect.deployments.deployments import run_deployment
-from prefect.logging import get_logger, get_run_logger
+from prefect.logging import get_run_logger
 from pydantic import BaseModel
-from vespa.io import VespaQueryResponse, VespaResponse
+from vespa.io import VespaResponse
 
+from flows.boundary import (
+    DocumentImporter,
+    DocumentImportId,
+    DocumentObjectUri,
+    TextBlockId,
+    convert_labelled_passage_to_concepts,
+    get_data_id_from_vespa_hit_id,
+    get_document_passage_from_vespa,
+    get_vespa_search_adapter,
+    load_labelled_passages_by_uri,
+    s3_obj_generator,
+    s3_object_write_text,
+    s3_paths_or_s3_prefixes,
+)
 from flows.inference import DOCUMENT_TARGET_PREFIX_DEFAULT
-from flows.utils import SlackNotify
+from flows.utils import SlackNotify, iterate_batch
 from scripts.cloud import (
     AwsEnv,
     ClassifierSpec,
@@ -35,7 +43,7 @@ from scripts.cloud import (
     get_prefect_job_variable,
 )
 from src.concept import Concept
-from src.exceptions import PartialUpdateError, QueryError
+from src.exceptions import PartialUpdateError
 from src.identifiers import WikibaseID
 from src.labelled_passage import LabelledPassage
 from src.span import Span
@@ -45,34 +53,6 @@ DEFAULT_INDEXING_TASK_BATCH_SIZE = 20
 HTTP_OK = 200
 CONCEPTS_COUNTS_PREFIX_DEFAULT: str = "concepts_counts"
 CONCEPT_COUNT_SEPARATOR: str = ":"
-
-# Needed to get document passages from Vespa
-# Example: CCLW.executive.1813.2418
-DocumentImportId: TypeAlias = str
-# Needed to load the inference results
-# Example: s3://cpr-sandbox-data-pipeline-cache/labelled_passages/Q787/v4/CCLW.executive.1813.2418.json
-DocumentObjectUri: TypeAlias = str
-# Passed to a self-sufficient flow run
-DocumentImporter: TypeAlias = tuple[DocumentImportId, DocumentObjectUri]
-
-
-class S3Accessor(BaseModel):
-    """Representing S3 paths and prefixes for accessing documents."""
-
-    paths: list[str] | None = None
-    prefixes: list[str] | None = None
-
-
-# AKA LabelledPassage
-TextBlockId: TypeAlias = str
-SpanId: TypeAlias = str
-# The ID used in Vespa, that we don't keep in our models in the CPR
-# SDK, that is in a Hit.
-VespaHitId: TypeAlias = str
-# The same as above, but without the schema
-VespaDataId: TypeAlias = str
-
-T = TypeVar("T")
 
 
 @dataclass()
@@ -252,7 +232,7 @@ def save_labelled_passages_by_uri(
         [labelled_passage.model_dump_json() for labelled_passage in labelled_passages]
     )
 
-    _ = _s3_object_write_text(
+    _ = s3_object_write_text(
         s3_uri=document_object_uri,
         text=object_json,
     )
@@ -284,114 +264,6 @@ def get_model_from_span(span: Span) -> str:
             f"Span has more than one labeller. Expected 1, got {len(span.labellers)}."
         )
     return span.labellers[0]
-
-
-def get_data_id_from_vespa_hit_id(hit_id: VespaHitId) -> VespaDataId:
-    """
-    Extract non-schema namespaced ID (last element after "::")
-
-    Example:
-    "CCLW.executive.10014.4470.623" from document passage ID like
-    "id:doc_search:document_passage::CCLW.executive.10014.4470.623".
-    """
-    splits = hit_id.split("::")
-    if len(splits) != 2:
-        raise ValueError(f"received {len(splits)} splits, when expecting 2: {splits}")
-    return splits[1]
-
-
-def get_document_passages_from_vespa(
-    document_import_id: DocumentImportId,
-    vespa_search_adapter: VespaSearchAdapter,
-) -> list[tuple[VespaHitId, VespaPassage]]:
-    """
-    Retrieve all the passages for a document in Vespa.
-
-    params:
-    - document_import_id: The document import id for a unique family document.
-    """
-    logger = get_logger()
-
-    logger.info(f"Getting document passages from Vespa: {document_import_id}")
-
-    vespa_query_response: VespaQueryResponse = vespa_search_adapter.client.query(
-        yql=(
-            # trunk-ignore(bandit/B608)
-            "select * from document_passage where family_document_ref contains "
-            f'"id:doc_search:family_document::{document_import_id}"'
-        )
-    )
-
-    if (status_code := vespa_query_response.get_status_code()) != HTTP_OK:
-        raise QueryError(status_code)
-
-    logger.info(
-        (
-            f"Vespa search response for document: {document_import_id} "
-            f"with {len(vespa_query_response.hits)} hits"
-        )
-    )
-
-    return [
-        (passage["id"], VespaPassage.model_validate(passage["fields"]))
-        for passage in vespa_query_response.hits
-    ]
-
-
-def get_document_passage_from_vespa(
-    text_block_id: str,
-    document_import_id: DocumentImportId,
-    vespa_search_adapter: VespaSearchAdapter,
-) -> tuple[VespaHitId, VespaPassage]:
-    """Retrieve a passage for a document in Vespa."""
-    logger = get_logger()
-
-    logger.info(
-        f"Getting document passage from Vespa: {document_import_id}, text block: {text_block_id}"
-    )
-
-    condition = qb.QueryField("family_document_ref").contains(
-        f"id:doc_search:family_document::{document_import_id}"
-    ) & qb.QueryField("text_block_id").contains(text_block_id)
-
-    yql = (
-        qb.select("*")  # pyright:ignore[reportAttributeAccessIssue]
-        .from_("document_passage")
-        .where(condition)
-    )
-
-    vespa_query_response: VespaQueryResponse = vespa_search_adapter.client.query(
-        yql=yql
-    )
-
-    if not vespa_query_response.is_successful():
-        raise QueryError(vespa_query_response.get_status_code())
-    if len(vespa_query_response.hits) != 1:
-        raise ValueError(
-            f"Expected 1 document passage for text block `{text_block_id}`, got {len(vespa_query_response.hits)}"
-        )
-
-    logger.info(
-        (
-            f"Vespa search response for document: {document_import_id} "
-            f"with {len(vespa_query_response.hits)} hits"
-        )
-    )
-
-    hit = vespa_query_response.hits[0]
-    passage_id = hit["id"]
-    passage = VespaPassage.model_validate(hit["fields"])
-
-    return passage_id, passage
-
-
-def load_labelled_passages_by_uri(
-    document_object_uri: DocumentObjectUri,
-) -> list[LabelledPassage]:
-    """Load and transforms the S3 object's body into LabelledPassages objects."""
-    object_json = json.loads(_s3_object_read_text(s3_path=document_object_uri))
-
-    return [LabelledPassage(**labelled_passage) for labelled_passage in object_json]
 
 
 def get_updated_passage_concepts(
@@ -651,7 +523,7 @@ def update_s3_with_some_successes(
 
     concepts_counts_uri = f"s3://{cache_bucket}/{concepts_counts_prefix}/{key_parts}"
 
-    _ = _s3_object_write_text(
+    _ = s3_object_write_text(
         s3_uri=concepts_counts_uri,
         text=serialised_concepts_counts,
     )
@@ -710,62 +582,6 @@ def get_parent_concepts_from_concept(
     )
 
     return parent_concepts, parent_concept_ids_flat
-
-
-def convert_labelled_passage_to_concepts(
-    labelled_passage: LabelledPassage,
-) -> list[VespaConcept]:
-    """
-    Convert a labelled passage to a list of VespaConcept objects and their text block ID.
-
-    The labelled passage contains a list of spans relating to concepts
-    that we must convert to VespaConcept objects.
-    """
-    logger = get_run_logger()
-
-    concepts: list[VespaConcept] = []
-
-    # The concept used to label the passage holds some information on the parent
-    # concepts and thus this is being used as a temporary solution for providing
-    # the relationship between concepts. This has the downside that it ties a
-    # labelled passage to a particular concept when in fact the Spans that a
-    # labelled passage has can be labelled by multiple concepts.
-    concept = Concept.model_validate(labelled_passage.metadata["concept"])
-    parent_concepts, parent_concept_ids_flat = get_parent_concepts_from_concept(
-        concept=concept
-    )
-
-    # This expands the list from `n` for `LabelledPassages` to `n` for `Spans`
-    for span_idx, span in enumerate(labelled_passage.spans):
-        if span.concept_id is None:
-            # Include the Span index since Span's don't have IDs
-            logger.error(
-                f"span concept ID is missing: LabelledPassage.id={labelled_passage.id}, Span index={span_idx}"
-            )
-            continue
-
-        if not span.timestamps:
-            logger.error(
-                f"span timestamps are missing: LabelledPassage.id={labelled_passage.id}, Span index={span_idx}"
-            )
-            continue
-
-        concepts.append(
-            VespaConcept(
-                id=span.concept_id,
-                name=concept.preferred_label,
-                parent_concepts=parent_concepts,
-                parent_concept_ids_flat=parent_concept_ids_flat,
-                model=get_model_from_span(span),
-                end=span.end_index,
-                start=span.start_index,
-                # These timestamps _should_ all be the same,
-                # but just in case, take the latest.
-                timestamp=max(span.timestamps),
-            )
-        )
-
-    return concepts
 
 
 async def partial_update_text_block(
@@ -830,41 +646,6 @@ async def run_partial_updates_of_concepts_for_batch(
             continue
 
 
-def get_vespa_search_adapter(
-    vespa_search_adapter: VespaSearchAdapter | None,
-) -> tuple[
-    ContextManager[str] | ContextManager[None],
-    VespaSearchAdapter,
-]:
-    """
-    Get a Vespa search adapter, if none provided.
-
-    It uses certs fetched, and then, saved to disk, if none was provided.
-    """
-    logger = get_run_logger()
-
-    # We want the directory used for the `VespaSearchAdapter` to be
-    # automatically cleaned up.
-    #
-    # To do this, we rely on the `tempfile.TemporaryDirectory`'s behaviour,
-    # or, a `contextlib.nullcontext` no-op, if a temporary directory
-    # wasn't needed.
-    if vespa_search_adapter is None:
-        logger.info("no Vespa search adapter, getting it from AWS secrets")
-        cm = tempfile.TemporaryDirectory()
-
-        vespa_search_adapter = get_vespa_search_adapter_from_aws_secrets(
-            cert_dir=cm.name,  # type: ignore
-            vespa_private_key_param_name="VESPA_PRIVATE_KEY_FULL_ACCESS",
-            vespa_public_cert_param_name="VESPA_PUBLIC_CERT_FULL_ACCESS",
-        )
-    else:
-        logger.info("Vespa search adapter provided")
-        cm = contextlib.nullcontext()
-
-    return cm, vespa_search_adapter
-
-
 async def run_partial_updates_of_concepts_for_batch_flow_or_deployment(
     documents_batch: list[DocumentImporter],
     documents_batch_num: int,
@@ -901,202 +682,6 @@ async def run_partial_updates_of_concepts_for_batch_flow_or_deployment(
         cache_bucket=cache_bucket,
         concepts_counts_prefix=concepts_counts_prefix,
     )
-
-
-def iterate_batch(
-    data: list[T] | Generator[T, None, None],
-    batch_size: int = DEFAULT_DOCUMENTS_BATCH_SIZE,
-) -> Generator[list[T], None, None]:
-    """Generate batches from a list or generator with a specified size."""
-    if isinstance(data, list):
-        # For lists, we can use list slicing
-        for i in range(0, len(data), batch_size):
-            yield data[i : i + batch_size]
-    else:
-        # For generators, accumulate items until we reach batch size
-        batch: list[T] = []
-        for item in data:
-            batch.append(item)
-            if len(batch) >= batch_size:
-                yield batch
-                batch = []
-        if batch:  # Don't forget to yield the last partial batch
-            yield batch
-
-
-def s3_paths_or_s3_prefixes(
-    classifier_specs: list[ClassifierSpec] | None,
-    document_ids: list[str] | None,
-    cache_bucket: str,
-    prefix: str,
-) -> S3Accessor:
-    """
-    Return the paths or prefix for the documents and classifiers.
-
-    - s3_prefix: The S3 prefix (directory) to yield objects from. Example:
-      "s3://cpr-sandbox-data-pipeline-cache/labelled_passages"
-    - s3_paths: A list of S3 object keys to yield objects from. Example:
-      [
-        "s3://cpr-sandbox-data-pipeline-cache/labelled_passages/Q787/v4/CCLW.executive.1813.2418.json",
-        "s3://cpr-sandbox-data-pipeline-cache/labelled_passages/Q787/v4/CCLW.legislative.10695.6015.json",
-      ]
-    """
-    logger = get_run_logger()
-
-    match (classifier_specs, document_ids):
-        case (None, None):
-            # Run on all documents, regardless of classifier
-            logger.info("run on all documents, regardless of classifier")
-            s3_prefix: str = "s3://" + os.path.join(
-                cache_bucket,
-                prefix,
-            )
-            return S3Accessor(paths=None, prefixes=[s3_prefix])
-
-        case (list(), None):
-            # Run on all documents, for the specified classifier
-            logger.info("run on all documents, for the specified classifier")
-            s3_prefixes = [
-                "s3://"
-                + os.path.join(
-                    cache_bucket,
-                    prefix,
-                    classifier_spec.name,
-                    classifier_spec.alias,
-                )
-                for classifier_spec in classifier_specs
-            ]
-            return S3Accessor(paths=None, prefixes=s3_prefixes)
-
-        case (list(), list()):
-            # Run on specified documents, for the specified classifier
-            logger.info("run on specified documents, for the specified classifier")
-            document_paths = [
-                "s3://"
-                + os.path.join(
-                    cache_bucket,
-                    prefix,
-                    classifier_spec.name,
-                    classifier_spec.alias,
-                    f"{doc_id}.json",
-                )
-                for classifier_spec in classifier_specs
-                for doc_id in document_ids
-            ]
-            return S3Accessor(paths=document_paths, prefixes=None)
-
-        case (None, list()):
-            raise ValueError(
-                "if document IDs are specified, a classifier "
-                "specifcation must also be specified, since they're "
-                "namespaced by classifiers (e.g. "
-                "`s3://cpr-sandbox-data-pipeline-cache/labelled_passages/Q787/"
-                "v4/CCLW.legislative.10695.6015.json`)"
-            )
-
-
-def _s3_object_write_text(s3_uri: str, text: str) -> None:
-    """Write text content to an S3 object."""
-    # Parse the S3 URI
-    s3_path: Path = Path(s3_uri)
-    if len(s3_path.parts) < 3:
-        raise ValueError(f"Invalid S3 path: {s3_path}")
-
-    bucket: str = s3_path.parts[1]
-    key = str(Path(*s3_path.parts[2:]))
-
-    # Create BytesIO buffer with the text content
-    body = BytesIO(text.encode("utf-8"))
-
-    # Upload to S3
-    s3 = boto3.client("s3")
-    _ = s3.put_object(Bucket=bucket, Key=key, Body=body, ContentType="application/json")
-
-
-def s3_obj_generator_from_s3_prefixes(
-    s3_prefixes: list[str],
-) -> Generator[
-    DocumentImporter,
-    None,
-    None,
-]:
-    """Return a generator that yields keys from a list of S3 prefixes."""
-    logger = get_logger()
-
-    for s3_prefix in s3_prefixes:
-        try:
-            # E.g. Path("s3://bucket/prefix/file.json").parts[1] == "bucket"
-            bucket = Path(s3_prefix).parts[1]
-            object_keys = _get_s3_keys_with_prefix(s3_prefix=s3_prefix)
-            for key in object_keys:
-                id: DocumentImportId = Path(key).stem
-                key: DocumentObjectUri = os.path.join("s3://", bucket, key)
-
-                yield id, key
-        except Exception as e:
-            logger.error(
-                f"failed to yield from S3 prefix. Error: {str(e)}",
-            )
-            continue
-
-
-def s3_obj_generator_from_s3_paths(
-    s3_paths: list[str],
-) -> Generator[
-    DocumentImporter,
-    None,
-    None,
-]:
-    """
-    Return a generator that yields keys from a list of S3 paths.
-
-    We extract the key from the S3 path by removing the first two
-    elements in the path.
-
-    E.g. "s3://bucket/prefix/file.json" -> "prefix/file.json"
-    """
-    logger = get_logger()
-    for s3_path in s3_paths:
-        try:
-            id: DocumentImportId = Path(s3_path).stem
-            uri: DocumentObjectUri = s3_path
-            yield id, uri
-        except Exception as e:
-            logger.error(
-                f"failed to yield from S3 path. Error: {str(e)}",
-            )
-            continue
-
-
-def s3_obj_generator(
-    s3_prefixes: list[str] | None,
-    s3_paths: list[str] | None,
-) -> Generator[
-    DocumentImporter,
-    None,
-    None,
-]:
-    """
-    Return a generator that returns S3 objects for each path or prefix.
-
-    These will be for each output from the inference stage, of
-    labelled passages.
-    """
-    logger = get_run_logger()
-
-    match (s3_prefixes, s3_paths):
-        case (list(), list()):
-            raise ValueError(
-                "Either s3_prefixes or s3_paths must be provided, not both."
-            )
-        case (list(), None):
-            logger.info("S3 object generator: prefixes")
-            return s3_obj_generator_from_s3_prefixes(s3_prefixes=s3_prefixes)
-        case (None, list()):
-            logger.info("S3 object generator: paths")
-            return s3_obj_generator_from_s3_paths(s3_paths=s3_paths)
-        case (None, None):
-            raise ValueError("Either s3_prefix or s3_paths must be provided.")
 
 
 @flow
