+++ tests/flows/test_count_family_document_concepts.py
@@ -21,7 +21,7 @@ async def test_load_parse_concepts_counts(
 ) -> None:
     """Test that we can load and parse concept counts from S3."""
     mock_concepts_counts_document_uri = sorted(mock_concepts_counts_document_keys)[0]
-    document_object_uri = f"s3://{mock_bucket}/{ mock_concepts_counts_document_uri }"
+    document_object_uri = f"s3://{mock_bucket}/{mock_concepts_counts_document_uri}"
 
     counter = await load_parse_concepts_counts(document_object_uri)
