+++ tests/flows/test_utils.py
@@ -3,6 +3,7 @@ import pytest
 from flows.utils import (
     SlackNotify,
     file_name_from_path,
+    iterate_batch,
     remove_translated_suffix,
 )
 
@@ -52,3 +53,21 @@ def test_remove_translated_suffix(file_name: str, expected: str) -> None:
     """Test that we can remove the translated suffix from a file name."""
 
     remove_translated_suffix(file_name) == expected
+
+
+@pytest.mark.parametrize(
+    "data, expected_lengths",
+    [
+        # Lists
+        (list(range(50)), [50]),
+        (list(range(850)), [400, 400, 50]),
+        ([], [0]),
+        # Generators
+        ((x for x in range(50)), [50]),
+        ((x for x in range(850)), [400, 400, 50]),
+        ((x for x in []), [0]),
+    ],
+)
+def test_iterate_batch(data, expected_lengths):
+    for batch, expected in zip(list(iterate_batch(data, 400)), expected_lengths):
+        assert len(batch) == expected
